pyproject.toml
[project]
name = "pyspart-test"
version = "0.1.0"
description = "Spark Test Environment for E-commerce KPI Demo"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy",
    "psycopg[binary]",
    "pyspark",
    "pydantic-settings",
    "alembic",
]

[project.optional-dependencies]
test = [
    "pytest",
    "httpx",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
